packagegroup: add test specific packagegroup
We initially started to add test packages in each individual image, but it is
becoming painful (and inefficient) to maintain. This patch create RPB test
packagegroups to better organize RPB test images.

The packagegroups are split in several packages. It is expected that all test
images will use packagegroup-rpb-tests which provides the all the packages
needed for 'console' images.

This change is a no-op, images are not modified.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>