packagegroup: split -rpb into multiple recipes
Commit 92f304e9133 (rpb*-image: create packagegroup to better organize image)
introduced a regression, since

* packagegroup-rpb recipe is required for -console image, to all packages from
packagegroup-rpb-x11 are built as well, though they are not included in the
image.

* the commit also adds Gstreamer in the -console image. When building -console
with 'x11' enabled, it will bring a lot of X packages regardless of
packagegroup-rpb-x11

As such, we are now doing splitting the packagegroup recipes into multiple
recipes to avoid building packages not needed when building only -console
image, and we remove Gstreamer from -console for now.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>