[rfc] piglit: work around errors on GBM and EGL.
More information can be found here:
  https://projects.linaro.org/browse/BB-418

First patch is about Piglit assuming GBM includes gbm_bo_map()
and gbm_bo_unmap() because of GBM's major version (>= 12.1),
but this turns out not to be the case with our Mali binary.

Second patch regards the lack of eglGetPlatformDisplay (and
other related symbols), which do exist as their EXT counter-
parts in Mali (eglGetPlatformDisplayEXT, etc.). This failure
is about testing eglGetPlatformDisplay, so the patch removes
the test entirely.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>