packagegroup-rpb-{x11,weston}: add kmscube
kmscube is a very convenient graphic test app, and it is always useful to have
it handy. Adding kmscube in packagegroup-rpb instead would bring mesa and
gstreamer in the headless console image which is probably not a good idea. so
adding it in the graphical images only.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>
(cherry picked from commit ef5def8de9d89505b9f512e71b91072840927e3a)