lkft-metadata bbclass: based on buildhistory with additional metadata
We need to expose some metadata on the QA reports. For a few recipes, we
started by using bitbake -e output but it doesn't scale. buildhistory
contains most of the information that we need except SRC_URI. The solution
is to introduce lkftmetadata class, which is based on buildhistory with the
additional metadata that we want to expose.

* based on buildhistory.bbclass
* replace all buildhistory occurrences by lkftmetadata
* add SRC_URI logging and separate it with newlines
* always create the parent directory

Signed-off-by: Koen Kooi <koen.kooi@linaro.org>