mesa-gl_%.bbappend: remove libgbm and libgbm-dev from PACKAGES
Using mesa-gl instead of mesa as libgl and mesa provider in the SGX case
is not enough.

I also had to add the bbappend to remove libgbm and libgbm-dev from mesa-gl's
PACKAGES as mesa-gl (with gbm removed from PACKAGECONFIG) creates zero-sized
libgbm and libgbm-dev packages which confuse libgbm's do_packagedata.

Signed-off-by: Andrey Konovalov <andrey.konovalov@linaro.org>