classes/insane: look for RPROVIDERs in package_qa
When an RDEPENDency is RPROVIDED, the insane bbclass will not
really get the packaged information for the package that is
actually providing said RDEPENDency. Since Bitbake has no
clue as to who will be providing those files (there might be
more than one RPROVIDEr for that RDEPENDency), an error is
reported due to missing RDEPENDencies.

Exempli gratia:
Package foo RPROVIDES libone.
Package bar RDEPENDS on libone. It builds and links against
libone.so.1()(64bits).
When package bar goes through package_qa, an error is thrown
because the list of files provided by "libone" (not a real
package) comes up empty, and therefore no provider for
libone.so.1()(64bits) can be found.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>