kselftests: pass --sysroot to Clang
Using --sysroot allows Clang to find headers installed by OE,
instead of the host's.

This is one example where the host's headers are being used
(and then missing):
  In file included from test_l4lb_noinline.c:5:
  In file included from /usr/include/string.h:25:
  /usr/include/features.h:364:12: fatal error: 'sys/cdefs.h' file not found
  #  include <sys/cdefs.h>
             ^
  1 error generated.

Furthermore, those headers that are found, might be found on
the host, instead of OpenEmbedded's sysroot.

Bitbake already does this with GCC, e.g. in run.do_compile:
  export CC="aarch64-linaro-linux-gcc --sysroot=/oe/build/tmp-rpb-glibc/sysroots/hikey"

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>