.github: initial GitHub Actions workflow for premerge test
Up until now, we were using Linaro CI (Jenkins) to test PR, this
commit changes to Github Actions and tuxsuite for Yocto.

We use pull_request_target event so that the workflow runs in the
context of the base branch, e.g. with 'safe' (merged) code to allow
testing PR from external repos.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>