rpb.inc: stop using ARM toolchain
There is little point in using ARM toolchain, the main (FSF) is pretty
good enough. Stop using the ARM toolchain and switch to the default so
that we do not diverge from main OE/Yocto and (potentially) don't have
to cope with toolchain/codegen bugs which are not present/tested/handled
by the rest of the community.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>